% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countNumbers}
\alias{countNumbers}
\title{Count genotype combinations at 2 SNPs}
\arguments{
\item{X}{numeric matrix of genotypes}
}
\value{
\code{count} vector of counts of 9 possible genotypes at SNP pair
}
\description{
Count genotype combinations at 2 SNPs
}
