% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daughterSire.R
\docType{data}
\name{daughterSire}
\alias{daughterSire}
\title{targetregion: allocation of paternal half-sib families}
\format{An object of class \code{integer} of length 265.}
\usage{
daughterSire
}
\description{
Vector of sire ID for each progeny
}
\keyword{datasets}
