% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{felsenstein}
\alias{felsenstein}
\title{Felsenstein's genetic map function}
\usage{
felsenstein(K, x, inverse = F)
}
\arguments{
\item{K}{parameter (numeric) corresponding to the intensity of crossover
interference}

\item{x}{vector of recombination rates}

\item{inverse}{logical, if FALSE recombination rate is mapped to Morgan unit,
if TRUE Morgan unit is mapped to recombination rate (default is FALSE)}
}
\value{
vector of genetic positions in Morgan units
}
\description{
Calculation of genetic distances from recombination rates given
  an interference parameter
}
\examples{
  felsenstein(0.1, seq(0, 0.5, 0.01))
}
\references{
Felsenstein, J. (1979) A mathematically tractable family of
  genetic mapping functions with different amounts of interference. Genetics
  91:769-775.
}
