% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.chr.R
\docType{data}
\name{genotype.chr}
\alias{genotype.chr}
\title{targetregion: progeny genotypes}
\format{An object of class \code{matrix} with 265 rows and 300 columns.}
\usage{
genotype.chr
}
\description{
matrix of progeny genotypes in target region on chromosome BTA1
}
\keyword{datasets}
