% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglikfun}
\alias{loglikfun}
\title{Calculate log-likelihood function}
\arguments{
\item{counts}{integer vector of observed 2-locus genotype}

\item{fAA}{frequency of maternal haplotype 1-1}

\item{fAB}{frequency of maternal haplotype 1-0}

\item{fBA}{frequency of maternal haplotype 0-1}

\item{fBB}{frequency of maternal haplotype 0-0}

\item{theta}{paternal recombination rate}
}
\value{
\code{lik} value of log likelihood at parameter estimates
}
\description{
Calculate log-likelihood function
}
