% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.chr.R
\docType{data}
\name{map.chr}
\alias{map.chr}
\title{targetregion: physical map}
\format{An object of class \code{data.frame} with 300 rows and 5 columns.}
\usage{
map.chr
}
\arguments{
\item{map.chr}{data frame
\describe{
  \item{SNP}{SNP index}
  \item{Chr}{chromosome of SNP}
  \item{locus_bp}{physical position of SNP in bp}
  \item{locus_Mb}{physical position of SNP in Mbp}
  \item{markername}{official SNP name}
}}
}
\description{
SNP marker map in target region on chromosome BTA1 according to
 ARS-UCD1.2
}
\keyword{datasets}
