% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapSire.R
\docType{data}
\name{hapSire}
\alias{hapSire}
\title{targetregion: sire haplotypes}
\format{An object of class \code{matrix} with 10 rows and 301 columns.}
\usage{
hapSire
}
\description{
matrix of sire haplotypes in target region on chromosome BTA1
}
\keyword{datasets}
